+++ src/Plugins/Shipping.ByWeight/Models/ShippingByWeightListModel.cs
@@ -33,7 +33,10 @@ namespace SmartStore.Plugin.Shipping.ByWeight.Models
         [SmartResourceDisplayName("Plugins.Shipping.ByWeight.Fields.ShippingChargeAmount")]
         public decimal AddShippingChargeAmount { get; set; }
 
-
+        [SmartResourceDisplayName("Plugins.Shipping.ByWeight.Fields.SmallQuantitySurcharge")]
+        public decimal SmallQuantitySurcharge { get; set; }
+        [SmartResourceDisplayName("Plugins.Shipping.ByWeight.Fields.SmallQuantityThreshold")]
+        public decimal SmallQuantityThreshold { get; set; }
 
         [SmartResourceDisplayName("Plugins.Shipping.ByWeight.Fields.LimitMethodsToCreated")]
         public bool LimitMethodsToCreated { get; set; }